requirements/requirements.txt
appdirs==1.4.4
asgiref==3.2.3
attrs==20.1.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Django==3.0.7
django-extensions==2.2.5
django-filter==2.2.0
django-kronos==1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
metadata-parser==0.10.0
pathspec==0.8.0
Pillow==7.0.0
psycopg2-binary==2.8.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.7.14
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.7
